Revert "Extract `Journey.Step` to a separate file and call it `JourneyStep`"
This reverts commit 0de09531a3181ae9c58d9bb80bffc84e22ce791a.
